Improved visualization, calculated efficiency
The draw-graph procedure now arranges turtles into appropriate
supply and demand curves (rather than a simple X). The equilibrium
price and quantity occur at the intersection of the two curves.
The quantity axis is technically the percentage of the longer of
the supply and demand schedules.

I've also added a reporter that calculates the maximum possible
total surplus so the model can calculate actual efficiency.